refactor : api.support.config -> api.support.api 로 디렉토리 이름 수정
